Add a first version of pixel sorter
For the moment it sorts with Hue, but this can be changed
by changing the index for comparison in the function
<Comparator> @l86. 1 will correspond to Saturation, and 2
will correspond to Brightness.

A good idea for further develpment includes finding a
subset in <sortedPixels> with the same hue value, and then
sort that subset using either Saturation, or Brithness, (and
repeat if necessary).

Also, it'd get better results if it sorts multiple Xs at a time,
instead of only one.

Try to implement a different drawing direction for the sorted
pixel array.